Add option to train non-pulse-width Arrow model
Currently, a number of elements are hard-coded in trainArrow.R,
such that it is impossible to train a nucleotide-only, non-pulse-width
model.  Even hard-coding the change, the process fails at a couple of
places based on the assumption that all models predict PulseWidth.
For Sequel related work this is fine, but this makes it a pain
to train and test non-Pw models for RSII chemistries.

This patch adds a new boolean flag to enable or disable the prediction
of Pulse-Width by the model, which also serves to trigger a handful of
fixes such that the training and plotting doesn't fail when disabled.