Get CCS plotting script moved over to use bh2Reporter API
This is a first try to get the CCSplotting script moved over to use bh2Reporter. Once the workflow is generated, I replace the old CCSplotting script with this updated one, and running run.sh will generate plots. Remain questions: 1, package feather is not available. 2, need to copy Bauhaus2.R to teh script folder. 3, need to create report folder to run the workflow.